Add URL rewrite
Add URL rewrite per tutorial.
According to Traversy, series of Apache directives.
Disables multiviews, so no confusion between /test and /test.php.
Turns on RewriteEngines.
Sets the rewrite base.
Sets two rewrite conditions, so that if the conditions do not obtain,
where the file is not found as described in the URL path,
the RewriteRule directs to the correct parameters off of index.php.