Create a Core class
Create a core class with protected variables, as well as two public methods:
one to echo the parameters after a url, one to run that function.
Instantiated the class on the index page.
Given the .htaccess modifications, this should mean that `root/test`
should behave just like `index.php?url=test`, with the test parameter
printing on the page.