Modify getUrl method and construct method
Modify getUrl method to trim the trailing slash after the parameters,
sanitize the url, and create an array based on the remaining slashes,
but only when there is such a parameter specified.

(No more errors when nothing is specified!)

When constructed, check to see if there is a controller file that
corresponds to the first part of the url.

If there is, make that currentController (instead of Pages).

Require the file corresponding to currentController. EG, Pages, if
there is no file that corresponds to the url parameter.

Make current controller a new instance of that controller.