Add method and parameter functionality
Check to see if the URL has a second part.
If so, check to see if there is a method within the controller that
corresponds to that second part.
If not, the default is the index method.

Make any remaining parts of the url into the params variable, or make
params an empty array if not.

Use call_user_func_array to call the currentMethod on the currentController
(defaulting to the index method on the Pages controller if unspecified)
with any specified params.

I understand this last part the least.